feat(core): add `Result<T,E>` type for operation `success` and `error` handling
The `type Result<T,E>` forms the backbone for being able to provide a simple but powerful typesaftey when working with methods and functions within TS codebases.

A succesful `Result` will always return a simple object `{status:'sucess', data:T}` decribing the output of a method or function.

Likewise it would also be able to return and allow for better `Errors` or *Exceptions* handling as it returns `{status:"error, errror:E}`

Signed-off-by: Fuzzy Bear <28299972+aFuzzyBear@users.noreply.github.com>